Add files via upload
Small change in the chromatin addition due to name changes in the metadata, I removed the cell_line merging requirement, as this is not needed and caused the merge to remove the KO cell lines (other cell line). 
These changes had some small impact on Figure 6 (and put the correct time label on the LBR and LMNA KO experiments).